fix cow rootmeta handle lifecycle
fix

fmt

fix

remove mgr
